\name{synPerturbed}
\alias{synPerturbed}
\docType{data}
\title{
  Synthetically Perturbed FCS.
}
\description{
  This is a FCS file in which a subset of one parameter was artificially
  perturbed so as to have a much higher fluorescent intensity than the
  remainder of the parameter's observations.
}
\examples{data(synPerturbed)}
\format{ A flowFrame with 17 observables and 76466 cells. }
\details{
  Cells during a specific time period had their fluorescent intensities
  increased on channel <V705-A>.
}
\keyword{datasets}
